+++ packages/go_router/lib/src/match.dart
@@ -325,6 +325,7 @@ class RouteMatch extends RouteMatchBase {
       name: route.name,
       path: route.path,
       extra: matches.extra,
+      topRoute: matches.last.route,
     );
   }
 }
@@ -382,6 +383,7 @@ class ShellRouteMatch extends RouteMatchBase {
       pathParameters: matches.pathParameters,
       pageKey: pageKey,
       extra: matches.extra,
+      topRoute: matches.last.route,
     );
   }